+++ src/pt/exhentainetbr/AndroidManifest.xml
@@ -2,7 +2,7 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android">
     <application>
         <activity
-            android:name=".pt.exhentai.ExHentaiUrlActivity"
+            android:name=".pt.exhentainetbr.ExHentaiNetBRUrlActivity"
             android:excludeFromRecents="true"
             android:exported="true"
             android:theme="@android:style/Theme.NoDisplay">
