+++ src/pt/exhentainetbr/res/mipmap-hdpi/ic_launcher.png
