+++ src/pt/exhentainetbr/res/mipmap-mdpi/ic_launcher.png
