+++ src/pt/exhentainetbr/res/mipmap-xhdpi/ic_launcher.png
